fix bug in user defined network detection
Unfortunately there is no sane option to check whether a
docker network is user defined, but we need that information
in order to decide whether we can carry over the ip addresses.
This commit adds a try/except block that tries to set the ip
addresses on the new container, and falls back to not setting
them when an exception occurs.

Signed-off-by: julian <dev@jneureuther.de>